% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgreemGenerTmp.R
\name{AgreemGeneralizedVector}
\alias{AgreemGeneralizedVector}
\title{Function to calculate the overall proportion agreement and specific agreement.}
\usage{
AgreemGeneralizedVector(df)
}
\arguments{
\item{df}{A dataframe or matrix of N x P with N the number of observations and P the number of raters.}
}
\value{
A list containing the overall proportion of agreement, the specific agreement, the number of raters, the number of observations
, the original dataframe and the (marginal) prevalences.
}
\description{
This function can be used in case of multinomial variables and when there are more than two raters. Note that it can also be used for
 dichotomous variables and when there are two raters
}
\note{
This is a temporary function! A more advanced function with more options and output will be made public as soon as the article is published.
}
\examples{
data(PsychMorbid)

AgreemGeneralizedVector(PsychMorbid)
}
