% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAplotMultipleR.R
\name{BAplotMultipleR}
\alias{BAplotMultipleR}
\title{Modified Bland-Altman plot for multiple raters}
\usage{
BAplotMultipleR(rater, subject, variable, data, LoA = c("Default", "loess"),
  ArgzLoess = list(span = 1, degree = 1), symbols.graph = NULL,
  colSymbols = NULL, Legend = T, ArgzLegend = list(x = "topright"),
  xlab = "Average", ylab = "Difference", ...)
}
\arguments{
\item{rater}{Variable indicating which rater made the rating.}

\item{subject}{Variable for the subject.}

\item{variable}{Variable containing the ratings of the subjects.}

\item{data}{The dataframe}

\item{LoA}{The limits of agreement that have to be used. See Details.}

\item{ArgzLoess}{Arguments for \code{loess.smooth} if \code{LoA = "loess"}.}

\item{symbols.graph}{Optional, the symbols to be used in the graph for the raters.}

\item{colSymbols}{Optional, the color of the symbols}

\item{Legend}{Logical, indicates whether a legend has to be printed. Default is \code{TRUE}.}

\item{ArgzLegend}{The arguments for the legend, see \code{\link{legend}}.}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}

\item{...}{Additional arguments to be passed to \code{\link{plot}}.}
}
\value{
Returns a list with the following objects

\item{AnovaSummary}{The summary of the two-way ANOVA}

\item{AvgPerSubject}{The average score per subject.}

\item{LoA}{The limits of agreement calculated according to Jones et al. (2011).}

\item{ICC}{The intraclass correlation coefficients. See \code{\link{ICC}}}
}
\description{
Modified Bland-Altman plot for multiple raters
}
\details{
The limits of agreement (LoA) have to be interpreted differently than in Bland-Altman plots with 2 raters. In the modified Bland-Altman
plots, the LoA indicate how different an individual rater can be when compared with the mean of all the raters (Jones et al., 2011). However, as with the
Bland-Altman plots for 2 raters, it may be that the variability of the differences increase as the magnitude of the measurement increases (Bland and
Altman, 1999). Alternative LoA can then be plotted using a method based on the method of Royston and Wright (1998). Specifying \code{LoA="loess"}, we get
approximate 2.5 and 97.5 percentile curves. In this method, the standard deviation of the difference scores per subject is calculated and the LoA
per subject are calculated as -/+ 1.96 times the standard deviation. A loess fit is then used to connect these LoA per subject. Note that we assume
that the difference scores are normally distributed (with mean 0 as we are working with centered values).
}
\examples{
data(DentalStudy)
BAplotMultipleR(dentist, patient, DMFS, DentalStudy)
}
\references{
Bland, J.M., Altman, D.G.(1999). Measuring agreement in method comparison studies. \emph{Statistical Methods In Medical Research},
Vol.8(2), pp. 135-160

Jones, M., Dobson, A., O'Brian, S. (2011). A graphical method for assessing agreement with the
mean between multiple observers using continuous measures. \emph{Int J Epidemiol}. Vol.40: pp. 1308-1313.

Royston, P., Wright, E.M. (1988). How to construct 'normal ranges' for fetal variables. \emph{Ultrasound Obstet Gynecol}, Vol.11: pp. 30-38.
}
