% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfintJack.R
\name{ConfintJack}
\alias{ConfintJack}
\title{Confidence intervals using the jackknife method.}
\usage{
ConfintJack(object, alpha = 0.05)
}
\arguments{
\item{object}{Object of type CohenK or FleissK, see \code{\link{CohenK}} and \code{\link{FleissK}}.}

\item{alpha}{Alpha-level for computing the confidence intervals}
}
\value{
Object with the statistic and its confidence intervals
}
\description{
This function can be used to calculate the confidence intervals of Cohen's Kappa or Fleiss' Kappa.
}
\examples{
data(PsychMorbid)
KappaStat = FleissK(PsychMorbid)
ConfintJack(KappaStat)
}
\references{
Fleiss J.L., Davies M. (1982). Jackknifing functions of multinomial frequencies, with an application to a
measure of concordance. \emph{Am J Epidemiol};115: 841-845.
}
\seealso{
\code{\link{CohenK}} and \code{\link{FleissK}}
}
