% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiagnosticAgreement.deVet.R
\name{DiagnosticAgreement.deVet}
\alias{DiagnosticAgreement.deVet}
\title{Calculating agreement for dichotomous variables.}
\usage{
DiagnosticAgreement.deVet(ratings, CI = T, ConfLevel = 0.95,
  correction = c("continuity", "Fleiss", "bootstrap"), NrBoot = 1000,
  Parallel = F, no_cores = detectCores() - 1)
}
\arguments{
\item{ratings}{A dataframe or matrix of N x P with N the number of observations and P the number of raters.}

\item{CI}{Logical, indicates if confidence intervals have to be calculated}

\item{ConfLevel}{The confidence level to be used in calculating the confidence intervals. Possible values
are \code{"continuity"}, \code{"Fleiss"} and \code{"bootstrap"}.}

\item{correction}{Method of calculating the confidence intervals (de Vet et al., 2017). The confidence intervals (CI) can be calculated using
continuity correction, Fleiss correction or by use of bootstrap samples.}

\item{NrBoot}{In case of bootstrap methodology to calculate the CI, the number of bootstrap samples.}

\item{Parallel}{Logical, indicates if parallel computing has to be used to compute the confidence intervals.
Implemented only when using bootstrapping to calculate the confidence intervals.}

\item{no_cores}{Number of cores if parallel computing is used. Default is 1 core less than the number of
cores present.}
}
\value{
A list with the following components:

\item{SumTable}{The summed table as defined in the article of de Vet et al. (2017)}

\item{ObservedAgreem}{The overall proportion of agreement (with CI).}

\item{SpecificAgreem}{The specific agreement for each of the categories (with CIs).}
}
\description{
Function to calculate specific agreement and overall proportion of agreement for dichotomous variables.
}
\details{
This function is based on the functions as given in the appendix of the article of de Vet et al. (2017).
}
\examples{
# Load data
data(Agreement_deVetArticle)
Df = Agreement_deVetArticle

DiagnosticAgreement.deVet(Df)
}
\references{
De Vet HCW, Mokkink LB, Terwee CB, Hoekstra OS, Knol DL. Clinicians are
right not to like Cohen’s k. \emph{BMJ} 2013;346:f2125.

de Vet, H.C.W., Terwee C.B., Knol, D.L., Bouter, L.M. (2006). When to use agreement
versus reliability measures. \emph{Journal of Clinical Epidemiology}, Vol.59(10), pp.1033-1039

de Vet, H.C.W., Dikmans, R.E., Eekhout, I. (2017). Specific agreement on dichotomous outcomes can be
calculated for more than two raters. \emph{Journal of Clinical Epidemiology}, Vol.83, pp.85-89
}
