% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FleissK.R
\name{FleissK}
\alias{FleissK}
\title{Fleiss' Kappa}
\usage{
FleissK(ratings, weights = "unweighted", conflev = 0.95, N = Inf)
}
\arguments{
\item{ratings}{A dataframe or matrix of N x P with N the number of observations and P the number of raters.}

\item{weights}{The weight matrix that has to be used to calculate the reliability. Default is \code{unweighted},}

\item{conflev}{Confidence level to calculate the confidence intervals of Fleiss' Kappa.}

\item{N}{Final population correction factor for the variance of Fleiss' Kappa (Gwet, 2008).}
}
\value{
\item{method}{Method that was used to calculate reliability}

\item{subjects}{Number of subjects in the dataframe}

\item{nraters}{Number of raters}

\item{value}{Value for Fleiss' Kappa}

\item{StdErr}{The standard error of the estimated Kappa value}

\item{stat.name}{The corresponding test statistic}

\item{statistic}{The value of the test statistic}

\item{p.value}{The p-value for the test}

\item{alpha}{Alpha-level used to calculate the confidence intervals}

\item{Po}{Overall proportion of agreement}

\item{Pe}{Proportion of agreement expected by chance}

\item{ratings}{The original dataframe with the ratings}

\item{WeightsMatrix}{The weights matrix used.}
}
\description{
Function to calculate Fleiss' Kappa (Fleiss, 1971).
}
\examples{
# Load data
data(PsychMorbid)

FleissK(PsychMorbid)
}
\references{
Fleiss, J. L. (1971). Measuring nominal scale agreement among many raters. \emph{Psychological
Bulletin}, Vol.76(5), pp.378-382

Gwet, K.L. (2008). Variance Estimation of Nominal-Scale Inter-Rater Reliability with Random
Selection of Raters \emph{Psychometrika}, Vol.73(3), pp.407-430
}
