% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KappaMatrix.R
\name{Kappa.Matrix}
\alias{Kappa.Matrix}
\title{Matrix of Kappa-type coefficients}
\usage{
Kappa.Matrix(df, FleissKappa = F, CI.est = F, N.bt = 1000, parallel = F,
  NrCores = detectCores - 1)
}
\arguments{
\item{df}{A dataframe or matrix of N x P with N the number of observations and P the number of raters.}

\item{FleissKappa}{Logical, indicates if Fleiss' Kappa has to be calculated.}

\item{CI.est}{Logical, indicates if confidence intervals using bootstrapping have to be calculated.}

\item{N.bt}{Number of bootstrap samples.}

\item{parallel}{Logical, indicates if parallel computing has to be used.}

\item{NrCores}{Number of cores if parallel computing is used. Default is 1 core less than the number of
cores present.}
}
\value{
A list containing the kappa-matrix and Fleiss' Kappa if asked for.
}
\description{
Function to compute a matrix of kappa-type coefficients to assess the reliability of nominal variables
(Roberts & McNamee, 1998).
}
\examples{
data(PsychMorbid)
Kappa.Matrix(PsychMorbid)
}
\references{
Roberts C., McNamee R. (1998). A matrix of kappa-type coefficients to assess the reliability of
nominal scales. \emph{Statistics in medicine}, Vol.17(4), pp.471-88
}
